+++ CHANGES.rst
@@ -6,7 +6,8 @@ Changelog
 
 - Add limit attachments validation. Can be configured with environment variable.
   [cekk]
-
+- Export also compiling date in csv.
+  [cekk]
 
 2.4.0 (2022-09-08)
 ------------------